parsing

+++ libft/ft_strlen.c
@@ -6,16 +6,16 @@
 /*   By: alukongo <alukongo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/24 13:10:17 by alukongo          #+#    #+#             */
-/*   Updated: 2021/12/02 16:47:25 by alukongo         ###   ########.fr       */
+/*   Updated: 2021/11/26 13:22:40 by alukongo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include<stdio.h>
 #include"libft.h"
 //good
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i] != '\0')
